refactor: move query params to filters model to be standard
